refactor(Peer-Group): Create component for managing peer groups in a period
Update dialog title text and some styling.
- #407
