+++ src/app/teacher/peer-group-grading.module.ts
@@ -3,6 +3,7 @@ import { PeerGroupAssignedWorkgroupsComponent } from '../../assets/wise5/classro
 import { PeerGroupDialogComponent } from '../../assets/wise5/classroomMonitor/classroomMonitorComponents/peer-group/peer-group-dialog/peer-group-dialog.component';
 import { PeerGroupGroupingComponent } from '../../assets/wise5/classroomMonitor/classroomMonitorComponents/peer-group/peer-group-grouping/peer-group-grouping.component';
 import { PeerGroupMoveWorkgroupConfirmDialogComponent } from '../../assets/wise5/classroomMonitor/classroomMonitorComponents/peer-group/peer-group-move-workgroup-confirm-dialog/peer-group-move-workgroup-confirm-dialog.component';
+import { PeerGroupPeriodComponent } from '../../assets/wise5/classroomMonitor/classroomMonitorComponents/peer-group/peer-group-period/peer-group-period.component';
 import { PeerGroupUnassignedWorkgroupsComponent } from '../../assets/wise5/classroomMonitor/classroomMonitorComponents/peer-group/peer-group-unassigned-workgroups/peer-group-unassigned-workgroups.component';
 import { PeerGroupWorkgroupComponent } from '../../assets/wise5/classroomMonitor/classroomMonitorComponents/peer-group/peer-group-workgroup/peer-group-workgroup.component';
 import { AngularJSModule } from '../common-hybrid-angular.module';
@@ -14,6 +15,7 @@ import { SelectPeriodModule } from './select-period.module';
     PeerGroupDialogComponent,
     PeerGroupGroupingComponent,
     PeerGroupMoveWorkgroupConfirmDialogComponent,
+    PeerGroupPeriodComponent,
     PeerGroupUnassignedWorkgroupsComponent,
     PeerGroupWorkgroupComponent
   ],
@@ -23,6 +25,7 @@ import { SelectPeriodModule } from './select-period.module';
     PeerGroupDialogComponent,
     PeerGroupGroupingComponent,
     PeerGroupMoveWorkgroupConfirmDialogComponent,
+    PeerGroupPeriodComponent,
     PeerGroupUnassignedWorkgroupsComponent,
     PeerGroupWorkgroupComponent
   ]
