+++ src/assets/wise5/classroomMonitor/classroomMonitorComponents/peer-group/peer-group-dialog/peer-group-dialog.component.html
@@ -1,27 +1,14 @@
-<h4 class="mat-body-2" i18n>Student Groupings for {{ componentLabel }} activity in Step {{ stepTitle }}</h4>
-<div>
+<h2 mat-dialog-title fxLayout="row wrap" fxLayoutAlign="start center" fxLayoutGap="8px">
+  <span i18n>Groupings for {{ componentLabel }} in {{ stepNumber }}</span>
+  <span fxFlex fxHide.xs></span>
   <select-period></select-period>
-  <div fxFlex></div>
-  <button mat-stroked-button
-      color="primary"
-      aria-label="Create a new group"
-      i18n-aria-label
-      (click)="createNewGroup()">
-    <mat-icon class="text-secondary" color="primary">group_add</mat-icon>
-    &nbsp;
-    <span i18n>New Group</span>
-  </button>
+</h2>
+<div mat-dialog-content class="dialog-content-scroll">
+    <peer-group-period *ngFor="let period of periods"
+        [nodeId]="nodeId"
+        [componentId]="componentId"
+        [period]="period"></peer-group-period>
 </div>
-<div cdkDropListGroup>
-  <peer-group-unassigned-workgroups
-      [unassignedWorkgroups]="unassignedWorkgroups"
-      (moveWorkgroup)="moveWorkgroup($event)">
-  </peer-group-unassigned-workgroups>
-  <peer-group-assigned-workgroups
-      [groupings]="groupings"
-      (moveWorkgroup)="moveWorkgroup($event)">
-  </peer-group-assigned-workgroups>
-</div>
-<div mat-dialog-actions class="close-div" fxLayoutAlign="end center">
+<div mat-dialog-actions fxLayoutAlign="end center">
   <button mat-button mat-dialog-close i18n>Close</button>
 </div>
