+++ src/assets/wise5/classroomMonitor/classroomMonitorComponents/peer-group/peer-group-period/peer-group-period.component.html
@@ -1,0 +1,22 @@
+<mat-card>
+  <div fxLayoutAlign="start center">
+    <h3 class="mat-display-1 accent-1" i18n>Period {{period.periodName}}</h3>
+    <span fxFlex></span>
+    <button mat-stroked-button
+        color="primary"
+        aria-label="Create a new group"
+        i18n-aria-label
+        (click)="createNewGroup()">
+      <mat-icon class="text-secondary" color="primary">group_add</mat-icon>
+      &nbsp;
+      <span i18n>New Group</span>
+    </button>
+  </div>
+  <div class="groups" cdkDropListGroup>
+    <peer-group-unassigned-workgroups [unassignedWorkgroups]="unassignedWorkgroups"
+        (moveWorkgroup)="moveWorkgroup($event)">
+    </peer-group-unassigned-workgroups>
+    <peer-group-assigned-workgroups [groupings]="groupings" (moveWorkgroup)="moveWorkgroup($event)">
+    </peer-group-assigned-workgroups>
+  </div>
+</mat-card>
