+++ src/assets/wise5/classroomMonitor/classroomMonitorComponents/peer-group/peer-group-period/peer-group-period.component.scss
@@ -1,0 +1,7 @@
+.mat-display-1 {
+  margin: 0;
+}
+
+.groups {
+  margin-top: 16px;
+}
